pkg/index: Retry whole file when EXIF data can't be read
When reading EXIF data from larger TIFF files, we might fail
to read the EXIF data when we only pass in the in-memory
prefix. This change identifies when the third-party library
encounters a short read on a tag/EXIF data and triggers a
retry with the whole file by returning an ErrUnexpectedEOF.

It also changes the third-party library error handling so
that we don't have to rely on the error message, but can use
the error value instead.

Change-Id: Ie5cdc1613db6ccac49d91a69827f11ca3406a74b